(#1335) support provisioning with no server config found
Previously a empty or corrupt server.conf was required
