+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 |Date      |Issue |Description                                                                                              |
 |----------|------|---------------------------------------------------------------------------------------------------------|
+|2021/08/09|1335  |Support entering provisioning mode when the supplied `server.conf` does not exist                        | 
 |2021/08/09|1333  |Disable RPC Auth during provisioning mode                                                                |
 |2021/08/04|1326  |Rename the `jetstream` adapter to `choria_streams`                                                       |
 |2021/08/03|1324  |Allow compiled-in Go agents to access the Submission system                                              |
